survey/migrations/0002_survey_template.py
# -*- coding: utf-8 -*-

from django.db import migrations, models


class Migration(migrations.Migration):

    dependencies = [
        ('survey', '0001_initial'),
    ]

    operations = [
        migrations.AddField(
            model_name='survey',
            name='template',
            field=models.CharField(max_length=255, null=True, blank=True),
        ),
    ]
