requirements_dev.txt
-r requirements.txt

# I18n
django-rosetta==0.9.0
# Lint
pylint==1.7.2
coverage==4.4.1
mock==2.0.0
coveralls==1.2.0
colorama==0.3.9
